+++ features/de.bund.bfr.knime.fsklab.feature/CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Change extension of generated R workspace files from .r to .RData. The URI of .RData is used in the manifest.
 
 ### Fixed
--  Tooltip of General information -> Creation date. https://github.com/SiLeBAT/FSK-Lab/issues/451 
+- Tooltip of General information -> Creation date. https://github.com/SiLeBAT/FSK-Lab/issues/451
+- Tooltip of General information -> Modification date. https://github.com/SiLeBAT/FSK-Lab/issues/450
 
 ## [1.8.1] - 2020-6-22